# log {long, short}: fetch branch information in parallel